Add script to parse log and plot job timeline
The script parses a "cluster_run.log" file to extract start/end times of
all jobs and generates a plot from it.  This can be useful to get an
overview of an ongoing run (e.g. to easily see if there are jobs that
take unusually long or if execution of new jobs is blocked for some
reason).

It is installed as entry point, so can be executed by calling

    cluster_utils_plot_timeline path/to/cluster_run.log