fix: Use helper class for json section generator
Passing the local function `find_json` as `section_generator` to the
`SectionFromJsonHook` causes a problem when the optimizer class (to
which the hook is added as argument) is pickled and later un-pickled in
a context where `find_json` is not available.

Instead use a globally defined helper class `StaticSectionGenerator` to
replace `find_json`.

Note: This is only a problem since we removed clouldpickle, which can
handle lambda functions (and was probably used for exactly this reason).
In this case it's easy enough to make it work with normal pickle,
though.