Save submission_hook_stats for offline report generation
Save `submission_hook_stats` to a pickle file at the end of each
iteration, so they can be loaded and used in `generate_report.py`.

Currently this only seems to include the git information, which doesn't
change between iterations, so it would technically be enough to save it
once in the beginning.  However, since this seems to be an extensible
thing, better directly put do it every iteration, so we can potentially
add more dynamic information.