Unify storage of report data and add metadata
Save a file metadata.json which contains relevant metadata of the run.
For now it contains the run type (hp_optimization/grid_search) and the
start time of the run.
This is needed for offline-generation of the report but might also be
useful for other purposes.

Use this metadata to determine the run type in `generate_report` instead
of guessing based on filenames.

Also unify the format in which the additional "report data" is stored.