Fix missing properly exiting jobs
When the results message arrives after the concluded message, the jobs
were previously counted as failed. This change adds a new job state,
CONCLUDED_WITHOUT_RESULTS, that is used in this case. When results
arrive in this state up to a timeout, the job is counted as successfully
concluded. If the timeout is reached, the job is counted as failed.