Disable black's preview style
`preview = true` enables experimental style changes of black and means
that with basically every new release, there is some small change in the
formatting.
By setting it to false, we stick to the "stable" style, which only gets
updated once at the beginning of every new year.  While this means, we
won't benefit from style improvements immediately, it reduces the amount
of failing CI checks and reformatting needed during the year.