Replace ClusterSystem.check_for_failure with .mark_failed_jobs
Change the structure to not call a function separately on each job but
instead use only one call that gets the whole list of jobs to check.
This enables to use only a single `sacct` call on Slurm, potentially
reducing the load on the system.

Add a method `Job.mark_failed` to aid developers implementing the
`mark_failed_jobs` method (mostly to ensure both status and error_info
are properly set).