Pin pandas to version 2.0.3 and remove dependency on jinja2
Pandas 2.1.2 has been released in the meantime, so depending on
`pandas[output-formatting]` should now also install jinja2 as part of
its dependencies.  However, 2.1 is not supporting Python 3.8 anymore.

So instead of using the latest version, pin to 2.0.3, which is the last
version that still supports Python 3.8 (so at least we are consistent
across Python versions).

Note that for this version the optional dependency group is
"output_formatting" (with underscore) and that still seems to work
there.  It probably needs to be changed back to a dash if we ever
upgrade the pandas version.