Only check for failed jobs once every minute on Slurm
We use the `sacct` command to check the status of jobs on Slurm.
However, this shouldn't be done too frequently to not put unnecessary
load on the system.  Therefore, keep track of the time the check was
performed the last time and only do it if enough time has passed since
then.