Re-enqueue jobs that stopped with exit_for_resume()
Implement the "exit for resume" feature in a cluster-independent way by
simply adding the jobs to the submission queue again.

This requires cluster system implementations to consider the value of
`job.waiting_for_resume`, e.g. to avoid recreation of already existing
run scripts.

Remove the resume handling from the local run script template and change
it such that output is appended to files instead of overwriting (so jobs
can be restarted without losing output of previous runs).

Don't re-enqueue jobs on Condor, as here we still rely on the cluster
taking care of it.

In `hp_optimization()` check for queued jobs in every iteration, not
only if a new job is to be added.  Otherwise re-enqueued jobs may not be
started and thus will never finish.