Add resume()/resume_fn() to ClusterSubmission
To avoid instance-checking in
CommunicationServer.handle_exit_for_resume(), add methods resume() and
resume_fn() to ClusterSubmission.

resume_fn() has a default implementation that simply re-enqueues the
job.  For Condor, this is overwritten with a method that does nothing.

resume() is a wrapper around resume_fn() that makes sure
job.waiting_for_resume is set (instead of doing this in
CommunicationServer).