chore: Build documentation and deploy to GitHub Pages
Rebuild is done for push to master, pull requests that have a change
inside docs/ and workflow_dispatch.
Publishing to Pages only happens for push on master or workflow_dispatch
if the corresponding option is set.

Pull request has to be handled in a separate workflow file because
setting permissions fails there (but is not needed when not publishing
anyways).