Refactor sacct output parsing and add unit test
Move the parsing of the output of sacct to a separate function, so it
can be unit-tested.

I also found out that by adding `-X` I can limit the output to only the
cumulative statistics of each job, omitting the intermediate results
(*.batch, etc.) which were skipped in the parsing anyway.