Use argparse in hp_optimization.py and grid_search.py
Use argparse to define the arguments passed to this script.  This gives
us automatic `--help/-h` arguments and makes it easier to extend in the
future.

So far, the arguments are still forwarded to `read_params_from_cmdline`
but it is a first step to getting rid of that function in the
user-called scripts.

Also move the related code to helper functions, so there is less code
duplication between grid_search and hp_optimization.

Related issue: #103