Use asyncio instead of pyuv
Python's asyncio provides everything we need to set up a simple UDP
server.  Use it instead of pyuv to get rid of an unnecessary third-party
dependency.

This also has the benefit that we avoid a direct dependency on a git
repository, which is blocking us from publishing cluster_utils to PyPI.