Rename read_params_from_cmdline and save_metrics_params
Rename
- `read_params_from_cmdline` to `initialize_job`, and
- `save_metrics_params` to `finalize_job`.

The purpose of the renaming is to better reflect when/why to call the
functions instead of focusing too much on technical details.

The old names are kept as aliases but are marked as deprecated.

BREAKING: I used this occasion to remove the following options from
`initialize_job/read_params_from_cmdline`:
- `make_immutable`: Parameters are always immutable now.  If one really
  needs a mutable structure, this is still possible by copying the data.
- `save_params`: It simply always saves now.