refactor: Add `base` sub-package
Add new sub-package `base` for everything that is used by both client
and server code.  This helps to better distinct which dependencies are
actually needed in each use-base.
As we want to keep dependencies for the client code minimal, this means
that also the base code should avoid third-party dependencies whenever
reasonably possible.

- Move MessageTypes to base sub-package.
- Move constants to base sub-package.  In practice most of the constants
  are only used in either client or server but splitting it up due to
  this seems a bit overkill to me.
- Move flatten_nested_string_dict to base.utils.  This is the only utils
  function that is used by both client and server.
- Move shared settings functions to base.settings.

With this change all internal imports in the `client` sub-package are
either from `client` itself or from `base`.  So all remaining top-level
modules are not used by the client code and can be moved to a dedicated
`server` sub-package in the next step.