Move remaining top-level modules to sub-package `server`
After recent refactorings, the remaining top-level modules are only
for the server processes, so move them all to a new sub-package
`server`.

Also remove the convenience imports of `grid_search` and
`hp_optimization` functions in top-level `__init__.py`.  They are not
supposed to be called by the regular user directly, so I don't think
they need to be there.  However, they were shadowing the modules of the
same name, which can be confusing.

The `grid_search`, `hp_optimization` and `parallel_execturo` modules are
still kept at top level for now, as they are called by the users via
`python -m cluster_utils.X`.