Handle shutdown outside of signal handler
Handling the shutdown on SIGINT (including sys.exit()) within the signal
handler function is problematic as the signal can be received multiple
times before the application actually exists.  This results in the
duplicated prints and potentially other issues.

Instead, only set a flag in the signal handler, which is then checked in
each iteration of the main loop.  Add a simple `SignalWatcher` utility
class for this.

Unfortunately this adds more redundant code in `hp_optimization()` and
`grid_search()` but I don't see an easy way to avoid that at the moment.