Do not use functions from `np.random`
The recommended way is not to create a random generator instance (e.g.
via `np.random.default_rng()` and use that.  It is supposed to be
significantly faster and statistically more robust.

In case we ever want to be able to set a seed for cluster_utils itself,
it's probably best to have one global generator used throughout the
package.  Add this, together with a little getter function `get_rng()`
in base/utils.py.

This fixes ruffs NPY warnings.