Support NumPy 2
Not too long ago, NumPy 2.0 got released.  cluster_utils didn't work
with it out of the box, so we restricted the dependency to numpy<2.

Looking into it now, it seems the only actual problem was in the
`Discrete` distribution when using booleans:  Using NumPy's `choice()`
converted the list of native bools to an array of np.bool.  Later one,
this clashed when passing the parameters to the job script, where they
are parsed with `ast.literal_eval()`.
Simply converting back to list with `.tolist()` seems to be enough to
fix this.