More useful error message when parsing the parameter dict fails
Show a more informative error message when the parameter dict, passed to
the job script, cannot be parsed by `ast.literal_eval`.  It now includes
the string to be parsed itself, to make debugging easier.

Since the string, that is passed here, is constructed by cluster_utils
when used normally, an error here most likely indicates a bug in
cluster_utils rather than a user mistake.