chore: Remove Workflow to build docs for GitHub Pages
Documentation is now hosted on Read the Docs.  The build there is
triggered through a webhook, so no need for a workflow anymore.

Note that the docs_pr.yml workflow is still kept as that is not
publishing but for verifying that the documentation can still be built
after modifications.