Tab-completion & history for interactive mode
Use the readline module to add tab-completion for the available
commands.  This also automagically gives us arrow-up-history for the
input prompt.

Note that the history is not perfect as it also includes input from
other prompts (e.g. the question whether an existing output directory
should be deleted) but I think it's good enough.

Resolves #47.