Do not exit command loop in error cases
Instead of just always exiting the command loop after one command, make
it based on the return value of the command function.  If all goes well,
it should return True to exit the loop but in case of an error (e.g.
invalid ID provided), it can return False, so the prompt stays open and
the user can try again.
It is always possible to exit the loop by simply pressing enter on an
empty line.